+++ packages/yourloops/pages/team/team-details-page.tsx
@@ -25,16 +25,17 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React, { useCallback, useRef } from "react";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useHistory, useParams } from "react-router-dom";
 
-import IconButton from "@material-ui/core/IconButton";
-import ArrowBackIcon from "@material-ui/icons/ArrowBack";
 import { makeStyles, Theme } from "@material-ui/core/styles";
-import GroupOutlinedIcon from "@material-ui/icons/GroupOutlined";
-import InfoOutlinedIcon from "@material-ui/icons/InfoOutlined";
+import ArrowBackIcon from "@material-ui/icons/ArrowBack";
+import Box from "@material-ui/core/Box";
 import DesktopMacIcon from "@material-ui/icons/DesktopMac";
+import GroupOutlinedIcon from "@material-ui/icons/GroupOutlined";
+import IconButton from "@material-ui/core/IconButton";
+import InfoOutlinedIcon from "@material-ui/icons/InfoOutlined";
 
 import { Team, useTeam } from "../../lib/team";
 import BasicDropdown from "../../components/dropdown/basic-dropdown";
@@ -46,36 +47,23 @@ import { commonTeamStyles } from "../../components/team/common";
 import { useAuth } from "../../lib/auth";
 
 const useStyles = makeStyles((theme: Theme) => ({
-  basicDropdown: {
-    marginLeft: "15px",
-    width: "180px",
-  },
-  body: {
-    display: "flex",
-    flexDirection: "row",
-  },
-  centerContent: {
-    justifyContent: "center",
+  activeLink: {
+    color: theme.palette.primary.main,
   },
   disableRipple: {
+    "paddingLeft": 0,
     "&:hover": {
       backgroundColor: "transparent",
     },
     "color": theme.palette.common.black,
   },
   drawer: {
-    minWidth: "370px",
-    maxWidth: "400px",
     position: "sticky",
     height: "200px",
     top: "80px",
   },
-  groupOutlinedIcon: {
-    margin: "0px",
-  },
   drawerTitle: {
     "display": "flex",
-    "flexDirection": "row",
     "alignItems": "center",
     "marginTop": "30px",
     "&:hover": {
@@ -85,30 +73,22 @@ const useStyles = makeStyles((theme: Theme) => ({
   refElement: {
     scrollMarginTop: "70px",
   },
-  root: {
-    display: "flex",
-    flexDirection: "row",
-  },
   separator: {
-    border: "1px solid #C4C4C4",
-    marginBottom: "50px",
-    marginTop: "50px",
+    border: `1px solid ${theme.palette.divider}`,
+    marginBottom: theme.spacing(6),
+    marginTop: theme.spacing(6),
   },
   teamDetails: {
-    maxWidth: "1030px",
+    maxWidth: "1140px",
+    padding: "0 20px",
   },
   teamInformation: {
     marginTop: "32px",
   },
-  teamSelection: {
-    display: "flex",
-    alignItems: "center",
-  },
   title: {
-    color: theme.palette.grey[800],
-    fontSize: "16px",
     fontWeight: 600,
     lineHeight: "20px",
+    textTransform: "uppercase",
   },
 }));
 
@@ -121,17 +101,18 @@ function TeamDetailPage(): JSX.Element {
   const authContext = useAuth();
   const { t } = useTranslation("yourloops");
   const { teamId } = paramHook as { teamId: string };
-  const [dropdownData, setDropdownData] = React.useState<{ selectedTeam: Team | null, teamNames: string[] } | null>(
+  const [dropdownData, setDropdownData] = useState<{ selectedTeam: Team | null, teamNames: string[] } | null>(
     { selectedTeam: null, teamNames: [] }
   );
+  const [activeLink, setActiveLink] = useState<string>("information");
   const isUserHcp = authContext.user?.isUserHcp();
-  const bodyClass = isUserHcp ? classes.body : `${classes.body} ${classes.centerContent}`;
 
   const teamInformation = useRef<HTMLDivElement>(null);
   const teamMembers = useRef<HTMLDivElement>(null);
   const teamAlarms = useRef<HTMLDivElement>(null);
 
   const scrollTo = (ref: React.RefObject<HTMLDivElement>) => {
+    setActiveLink(ref.current?.dataset.link as string);
     ref.current?.scrollIntoView({ behavior: "smooth" });
   };
 
@@ -142,7 +123,7 @@ function TeamDetailPage(): JSX.Element {
     });
   }, [getTeam, teamId, getMedicalTeams]);
 
-  React.useEffect(() => {
+  useEffect(() => {
     refresh();
   }, [refresh]);
 
@@ -159,80 +140,93 @@ function TeamDetailPage(): JSX.Element {
     <React.Fragment>
       {dropdownData?.selectedTeam &&
         <div>
-          <div className={classes.teamSelection}>
-            <IconButton className={classes.disableRipple} disableTouchRipple={true} onClick={redirectToDashboard}>
+          <Box display="flex" alignItems="center">
+            <IconButton className={classes.disableRipple} onClick={redirectToDashboard}>
               <ArrowBackIcon />
             </IconButton>
-            <GroupOutlinedIcon className={classes.groupOutlinedIcon} />
-            {t("team")}
-            <div className={classes.basicDropdown}>
-              <BasicDropdown
-                key={dropdownData.selectedTeam.name}
-                id={"team-basic-dropdown"}
-                defaultValue={dropdownData.selectedTeam.name}
-                values={dropdownData.teamNames}
-                onSelect={redirectToTeam} />
-            </div>
-          </div>
-          <div className={bodyClass}>
+            <GroupOutlinedIcon />
+            <Box marginLeft={0.5} marginRight={2}>{t("team")}</Box>
+            <BasicDropdown
+              key={dropdownData.selectedTeam.name}
+              id={"team-basic-dropdown"}
+              defaultValue={dropdownData.selectedTeam.name}
+              values={dropdownData.teamNames}
+              onSelect={redirectToTeam}
+            />
+          </Box>
+          <Box display="flex" justifyContent={isUserHcp}>
             {isUserHcp &&
               <div className={classes.drawer}>
                 <div
                   role="link"
-                  className={classes.drawerTitle}
+                  className={`${classes.drawerTitle} ${activeLink === "information" ? classes.activeLink : ""}`}
                   tabIndex={0}
                   onKeyDown={() => scrollTo(teamInformation)}
                   onClick={() => scrollTo(teamInformation)}
                 >
                   <InfoOutlinedIcon className={commonTeamClasses.icon} />
                   <Typography className={classes.title}>
-                    {t("information").toUpperCase()}
+                    {t("information")}
                   </Typography>
                 </div>
                 <div
                   role="link"
-                  className={classes.drawerTitle}
+                  className={`${classes.drawerTitle} ${activeLink === "members" ? classes.activeLink : ""}`}
                   tabIndex={0}
                   onClick={() => scrollTo(teamMembers)}
                   onKeyDown={() => scrollTo(teamMembers)}
                 >
                   <GroupOutlinedIcon className={commonTeamClasses.icon} />
                   <Typography className={classes.title}>
-                    {t("members").toUpperCase()}
+                    {t("members")}
                   </Typography>
                 </div>
                 <div
                   role="link"
-                  className={classes.drawerTitle}
+                  className={`${classes.drawerTitle} ${activeLink === "configuration" ? classes.activeLink : ""}`}
                   tabIndex={0}
                   onClick={() => scrollTo(teamAlarms)}
                   onKeyDown={() => scrollTo(teamAlarms)}
                 >
                   <DesktopMacIcon className={commonTeamClasses.icon} />
                   <Typography className={classes.title}>
-                    {t("telemonitoring-alarms").toUpperCase()}
+                    {t("events-configuration")}
                   </Typography>
                 </div>
               </div>
             }
-            <div className={classes.teamDetails}>
-              <div ref={teamInformation} className={`${classes.teamInformation} ${classes.refElement}`}>
-                <TeamInformation team={dropdownData.selectedTeam} refreshParent={refresh} />
-              </div>
-              {isUserHcp &&
-                <div>
-                  <div className={classes.separator} />
-                  <div ref={teamMembers} className={classes.refElement}>
-                    <TeamMembers team={dropdownData.selectedTeam} refreshParent={refresh} />
-                  </div>
-                  <div className={classes.separator} />
-                  <div ref={teamAlarms} className={classes.refElement}>
-                    <TeamAlarms team={dropdownData.selectedTeam} />
-                  </div>
+            <Box display="flex" justifyContent="center" margin="auto">
+              <div className={classes.teamDetails}>
+                <div
+                  ref={teamInformation}
+                  data-link="information"
+                  className={`${classes.teamInformation} ${classes.refElement}`}
+                >
+                  <TeamInformation team={dropdownData.selectedTeam} refreshParent={refresh} />
                 </div>
-              }
-            </div>
-          </div>
+                {isUserHcp &&
+                  <div>
+                    <div className={classes.separator} />
+                    <div
+                      ref={teamMembers}
+                      data-link="members"
+                      className={classes.refElement}
+                    >
+                      <TeamMembers team={dropdownData.selectedTeam} refreshParent={refresh} />
+                    </div>
+                    <div className={classes.separator} />
+                    <div
+                      ref={teamAlarms}
+                      data-link="configuration"
+                      className={classes.refElement}
+                    >
+                      <TeamAlarms />
+                    </div>
+                  </div>
+                }
+              </div>
+            </Box>
+          </Box>
         </div>
       }
     </React.Fragment>